rpc: finished up shim work for the rpc bridge
manager: changed to implementing ManagerService and using a shim

previously implemented the twirp server side directly, now implements
ManagerService instead and uses a twirp shim to do the rest.

streamer: changed to implementing StreamerService and using a shim

moved requests.go into api.go file and adjusted it for the type changes

irc: changed to implementing AnnounceService and using a shim

changed commands touching *Service endpoints to the new types if
required

config: TwirpClient's renamed to Client's

radio: implemented a Copy on the Status type, mainly for use in the
manager code

+++ database/track.go
@@ -233,7 +233,7 @@ func UpdateTrackPlayTime(h Handler, id radio.TrackID) error {
 // InsertPlayedSong inserts a row into the eplay table with the arguments given
 //
 // ldiff can be nil to indicate no listener data was available
-func InsertPlayedSong(h Handler, id radio.SongID, ldiff *int64) error {
+func InsertPlayedSong(h Handler, id radio.SongID, ldiff *int) error {
 	var query = `INSERT INTO eplay (isong, ldiff) VALUES (?, ?);`
 
 	_, err := h.Exec(query, id, ldiff)
